fix: add autoload-dev for Tests namespace and fix UrlHelperTest namespace
- Add autoload-dev section to composer.json for Tests namespace
- Rename tests/helpers to tests/Helpers for PSR-4 compliance
- Add proper namespace to UrlHelperTest.php

This fixes PHPUnit test discovery - previously only 6 tests were running
instead of ~132 due to missing namespace autoloading.
